Dev 2017.2 zedboard (#7)
* This commit includes the scripts for the flow with the ZedBoard and minor changes to the organization of the files.

The scripts for Vivado 2017.2 were moved to a dedicated directory and should be kept there. Upcoming scripts for Vivado 2019.2 will have their own folder.

The use of the "pack FGPU feature" has also been tested and shown to be working. Therefore, the FGPU_IP directory that was shipped together was removed. The scripts for manually packing the IP should be used instead.

To use this version of the flow, simply run main_implement.tcl. This script runs all the steps from IP packing to IP generation.

* Update README.md

* minor fix